# [#19288] Replace old white error screen by less intrusive error messages # [#19288] Fix regression: Bad category permissions in sample data
git-svn-id: file:///Users/oratzes/svn/kunena@2557 daf2aa89-cb5b-0410-b8ee-fe93ebc9b27d

+++ trunk/1.6/administrator/components/com_kunena/install/kunena.special.upgrade.1.0.8.php
@@ -20,37 +20,46 @@ $kunena_db = JFactory::getDBO ();
 $kunena_db->setQuery ( "CREATE TABLE #__kunena_favorites_temp SELECT thread, userid FROM #__kunena_favorites WHERE userid>0 GROUP BY thread, userid" );
 if ($kunena_db->query () == FALSE) {
 	$temporary = 0;
-	//	check_dbwarning("Unable to fix fb_favorites table. All Favorites will be removed.");
 }
 $kunena_db->setQuery ( "TRUNCATE #__kunena_favorites" );
 $kunena_db->query ();
 $kunena_db->setQuery ( "ALTER TABLE `#__kunena_favorites` DROP INDEX `thread`" );
 $kunena_db->query ();
 $kunena_db->setQuery ( "ALTER TABLE `#__kunena_favorites` ADD UNIQUE `thread`(`thread`,`userid`)" );
-$kunena_db->query (); // or check_dberror("Unable to alter fb_favorites table, please contact Kunena team at www.kunena.com!");
+$kunena_db->query ();
+if ($this->db->getErrorNum ())
+	throw new KunenaInstallerException ( $this->db->getErrorMsg (), $this->db->getErrorNum () );
 if ($temporary) {
 	$kunena_db->setQuery ( "INSERT INTO #__kunena_favorites (thread,userid) SELECT thread, userid FROM #__kunena_favorites_temp" );
-	$kunena_db->query (); //  or check_dbwarning("Unable to fix fb_favorites table. All Favorites will be removed.");
+	$kunena_db->query ();
+	if ($this->db->getErrorNum ())
+		throw new KunenaInstallerException ( $this->db->getErrorMsg (), $this->db->getErrorNum () );
 	$kunena_db->setQuery ( "DROP TABLE #__kunena_favorites_temp" );
-	$kunena_db->query (); // or check_dbwarning("Unable to remove temporary table (#__kunena_favorites_temp).");
+	$kunena_db->query ();
+	if ($this->db->getErrorNum ())
+		throw new KunenaInstallerException ( $this->db->getErrorMsg (), $this->db->getErrorNum () );
 }
 
 $temporary = 1;
 $kunena_db->setQuery ( "CREATE TABLE #__kunena_subscriptions_temp SELECT thread, userid, future1 FROM #__kunena_subscriptions WHERE userid>0 GROUP BY thread, userid" );
 if ($kunena_db->query () == FALSE) {
 	$temporary = 0;
-	//	check_dbwarning("Unable to fix fb_subscriptions table. All Subscriptions will be removed.");
 }
 $kunena_db->setQuery ( "TRUNCATE #__kunena_subscriptions" );
 $kunena_db->query ();
 $kunena_db->setQuery ( "ALTER TABLE `#__kunena_subscriptions` DROP INDEX `thread`" );
 $kunena_db->query ();
 $kunena_db->setQuery ( "ALTER TABLE `#__kunena_subscriptions` ADD UNIQUE `thread`(`thread`,`userid`)" );
-$kunena_db->query (); // or check_dberror("Unable to alter fb_subscriptions table, please contact Kunena team at www.kunena.com!");
+$kunena_db->query ();
+if ($this->db->getErrorNum ())
+	throw new KunenaInstallerException ( $this->db->getErrorMsg (), $this->db->getErrorNum () );
 if ($temporary) {
 	$kunena_db->setQuery ( "INSERT INTO #__kunena_subscriptions (thread,userid,future1) SELECT thread, userid, future1 FROM #__kunena_subscriptions_temp" );
-	$kunena_db->query (); // or check_dbwarning("Unable to fix fb_subscriptions table. All Subscriptions will be removed.");
+	$kunena_db->query ();
+	if ($this->db->getErrorNum ())
+		throw new KunenaInstallerException ( $this->db->getErrorMsg (), $this->db->getErrorNum () );
 	$kunena_db->setQuery ( "DROP TABLE #__kunena_subscriptions_temp" );
-	$kunena_db->query (); // or check_dbwarning("Unable to remove temporary table (#__kunena_subscriptions_temp).");
+	$kunena_db->query ();
+	if ($this->db->getErrorNum ())
+		throw new KunenaInstallerException ( $this->db->getErrorMsg (), $this->db->getErrorNum () );
 }
-
